pyproject.toml
[project]
name = "personal-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "fire>=0.7.0",
    "pillow>=11.2.1",
]
